ReefAngel_Wifi.cpp - code fixup, change int to byte
Changed a loop through the expansion relays to use a byte variable instead of an int variable to reduce the amount of storage used.

+++ ReefAngel_Wifi/ReefAngel_Wifi.cpp
@@ -246,7 +246,7 @@ void processHTTP()
 				s += intlength(ReefAngel.Relay.RelayMaskOff);
 #ifdef RelayExp
 				s += 296;
-				for ( int EID = 0; EID < MAX_RELAY_EXPANSION_MODULES; EID++ )
+				for ( byte EID = 0; EID < MAX_RELAY_EXPANSION_MODULES; EID++ )
 				{
 					s += intlength(ReefAngel.Relay.RelayDataE[EID]);
 					s += intlength(ReefAngel.Relay.RelayMaskOnE[EID]);
